Fix realtime/linuxcnc script installation
Unifies installation of realtime/linuxcnc script in the same
directory as expected by the configure.ac script.

Fixes regression introduced on #1326 which causes that 'realtime'
script is being installed in directory referred by ${libdir}
while the configure.ac expected it to be installed on 'lib' directory.
On systems where ${libdir} != 'lib' this generates discrepancy.

Also fixes proper $(prefix)/lib/linuxcnc directory creation.
When this was not created it caused 'realtime' script to be installed
as $(prefix)/lib/linuxcnc instaed of $(prefix)/lib/linuxcnc/realtime.